fix: SQL 치명적인 버그

+++ src/main/java/com/dku/council/domain/user/model/dto/response/ResponseUserInfoDto.java
@@ -14,8 +14,8 @@ public class ResponseUserInfoDto {
     private final String major;
     private final String department;
     private final String phoneNumber;
-    private final Integer writeCount;
-    private final Integer commentCount;
-    private final Integer likeCount;
+    private final Long writeCount;
+    private final Long commentCount;
+    private final Long likeCount;
     private final boolean admin;
 }
